+++ CHANGES.rst
@@ -10,6 +10,8 @@ Unreleased
 
 * Added cratedb_cluster_last_user_activity metric to the sql exporter
 
+* Fixed success notifications being sent too soon for update operations.
+
 2.15.0 (2022-09-28)
 -------------------